+++ src/components/CandleStickChart.js
@@ -1,0 +1,77 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+import { ChartCanvas, Chart } from "react-stockcharts";
+import { CandlestickSeries } from "react-stockcharts/lib/series";
+import { XAxis, YAxis } from "react-stockcharts/lib/axes";
+
+import { EdgeIndicator } from "react-stockcharts/lib/coordinates";
+
+import { discontinuousTimeScaleProvider } from "react-stockcharts/lib/scale";
+import { fitWidth } from "react-stockcharts/lib/helper";
+import { last } from "react-stockcharts/lib/utils";
+
+class CandleStickStockScaleChart extends React.Component {
+  render() {
+    const { type, data: initialData, width, ratio } = this.props;
+
+    const xScaleProvider = discontinuousTimeScaleProvider.inputDateAccessor(
+      (d) => d.date
+    );
+    const { data, xScale, xAccessor, displayXAccessor } =
+      xScaleProvider(initialData);
+    const xExtents = [
+      xAccessor(last(data)),
+      xAccessor(data[data.length - 100]),
+    ];
+
+    return (
+      <ChartCanvas
+        height={400}
+        ratio={ratio}
+        width={width}
+        margin={{ left: 50, right: 50, top: 10, bottom: 30 }}
+        type={type}
+        seriesName="MSFT"
+        data={data}
+        xScale={xScale}
+        xAccessor={xAccessor}
+        displayXAccessor={displayXAccessor}
+        xExtents={xExtents}
+      >
+        <Chart id={1} yExtents={(d) => [d.high, d.low]}>
+          <XAxis axisAt="bottom" orient="bottom" ticks={6} />
+          <YAxis axisAt="left" orient="left" ticks={5} />
+          <CandlestickSeries />
+
+          <EdgeIndicator
+            itemType="last"
+            orient="right"
+            edgeAt="right"
+            yAccessor={(d) => d.close}
+            fill={(d) => (d.close > d.open ? "#A2F5BF" : "#F9ACAA")}
+            stroke={(d) => (d.close > d.open ? "#0B4228" : "#6A1B19")}
+            textFill={(d) => (d.close > d.open ? "#0B4228" : "#420806")}
+            strokeOpacity={1}
+            strokeWidth={3}
+            arrowWidth={2}
+          />
+        </Chart>
+      </ChartCanvas>
+    );
+  }
+}
+
+CandleStickStockScaleChart.propTypes = {
+  data: PropTypes.array.isRequired,
+  width: PropTypes.number.isRequired,
+  ratio: PropTypes.number.isRequired,
+  type: PropTypes.oneOf(["svg", "hybrid"]).isRequired,
+};
+
+CandleStickStockScaleChart.defaultProps = {
+  type: "svg",
+};
+CandleStickStockScaleChart = fitWidth(CandleStickStockScaleChart);
+
+export default CandleStickStockScaleChart;
