+++ src/utils/utils.js
@@ -5,6 +5,46 @@ export const ws = new WebSocket(
   "wss://ws.binaryws.com/websockets/v3?app_id=" + app_id
 );
 
+export const candleInterval = {
+  one_minute: 60,
+  one_hour: 3600,
+  one_day: 86400,
+};
+
+export const closeStream = (stream_id) => {
+  console.log("closing ", stream_id);
+  ws.send(
+    JSON.stringify({
+      forget: stream_id,
+    })
+  );
+};
+
+export const getHistoricalData = (symbol, style, interval) => {
+  ws.send(
+    JSON.stringify({
+      ticks_history: symbol,
+      adjust_start_time: 1,
+      count: 100,
+      end: "latest",
+      start: 1,
+      style: style,
+      granularity: interval,
+    })
+  );
+
+  console.log("historical data req sent");
+};
+
+export const subscribeTickStream = (symbol) => {
+  ws.send(
+    JSON.stringify({
+      ticks: symbol,
+      subscribe: 1,
+    })
+  );
+};
+
 export const create_candle_data = (tick) => {
   let single_candle = {};
